Started working on the algorithm aspect
added a sensor for all four directions to detect a barrier on each side (update neighbors function and a bit at the end (event.type that I still have to figure out the purpose of. This is ending up to be a hard task for me since I can't remember the algorithms that I had revised before starting this project. Will need to revise on those soon...